+++ tests/test_observer.py
@@ -47,6 +47,7 @@ class BaseBinding(object):
         self.single_shot(0, action_trigger)
         self.single_shot(5000, partial(self.quit_mainloop, mainloop))
         self.run_mainloop(mainloop)
+        self.cleanup(observer)
         return event_slot, action_slots
 
 
@@ -86,8 +87,66 @@ class Qt4Binding(BaseBinding):
     def single_shot(self, timeout, callback):
         self.qtcore.QTimer.singleShot(timeout, callback)
 
+    def cleanup(self, observer):
+        pass
 
-BINDINGS = [Qt4Binding('PyQt4'), Qt4Binding('PySide')]
+
+class GlibBinding(BaseBinding):
+    name = 'Glib'
+
+    def __init__(self):
+        self.event_sources = []
+
+    @property
+    def observer_class(self):
+        return self._import_binding().GUDevMonitorObserver
+
+    def import_or_skip(self):
+        self.glib = py.test.importorskip('glib')
+        py.test.importorskip('gobject')
+
+    def create_mainloop(self):
+        return self.glib.MainLoop()
+
+    def quit_mainloop(self, mainloop):
+        mainloop.quit()
+
+    def run_mainloop(self, mainloop):
+        mainloop.run()
+
+    def wrap_callback(self, callback):
+        """
+        Remove the emitter from callback invocation, it is not of any
+        interest to the test code.
+        """
+        def _wrapper(obj, *args, **kwargs):
+            return callback(*args, **kwargs)
+        return _wrapper
+
+    def connect_signals(self, observer, event_slot, action_slots):
+        observer.connect('device-event', self.wrap_callback(event_slot))
+        signal_map = {'add': 'device-added', 'remove': 'device-removed',
+                      'change': 'device-changed', 'move': 'device-moved'}
+        for action, slot in action_slots.items():
+            observer.connect(signal_map[action], self.wrap_callback(slot))
+
+    def single_shot(self, timeout, callback):
+        def _callback():
+            callback()
+            return False
+        self.event_sources.append(self.glib.timeout_add(timeout, _callback))
+
+    def cleanup(self, observer):
+        # cleanup all event sources.  If this was not done, all tests save
+        # the first would spin endlessly.  Apparently older event sources
+        # kind of "confuse" the main loop of the current test
+        for source in self.event_sources:
+            self.glib.source_remove(source)
+        self.glib.source_remove(observer.event_source)
+        self.event_sources = []
+
+
+BINDINGS = [Qt4Binding('PyQt4'), Qt4Binding('PySide'), GlibBinding()]
 ACTIONS = ('add', 'remove', 'change', 'move')